Fixing context path bug in JIRA service hook
There was a bug that prevented certain urls from working
with the JIRA gem. Specifically, it looks like any url
with a non-empty context path didn't work. This commit
adds code to parse out the context path from each url and
instantiate the JIRA gem client using the parsed context
path.

Tests were added to check that different context paths
are parsed properly.

The version of the crashlytics-service gem was bumped
to 3.5.11.