Handle scenario when username != repo owner
Previously, the Bitbucket service hook made the assumption that
the username and repo owner were always the same. This commit
adds a repo owner field to handle situations when this assumption
is wrong. If the repo owner field is missing, the hook should
fall back to using the username. Users that verified their
accounts using the previous version should not be affected.